from_vec: return None
The previous error case of returning the original Vec makes the API
a bit awkward.

Rust's `?` operator can't be used without implementing From on a Vec:

    Vec1::from_vec(my_vec)?

I can see how this API came to be: by moving the variable in to the
function, you can't get it back in the error case. However, because
the Vec is empty, there is no reason to want it back anyway.