Fixed bug in splice where it didn't panic on *some* out of bounds cases.
(The cases where when the range was out of bounds but still did covered
the while slice (e.g. end > len) and the replacement iterator was empty,
in which case it did return a Size0Error instead of panicin which is
wrong because it's out of bounds and as such will never be reduced to
size 0).